moved contact data models to its own file
